Enhancement: Replace concrete class with an anonymous class
