+++ src/Tuple.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Spatie\Typed;
 
+use Iterator;
 use ArrayAccess;
 
 class Tuple implements ArrayAccess
@@ -23,7 +24,7 @@ class Tuple implements ArrayAccess
 
     public function set(array $data): self
     {
-        $iterator = new TupleIterator($this->types, $data);
+        $iterator = $this->createIterator($data);
 
         foreach ($iterator as $key => ['type' => $type, 'value' => $value]) {
             $data[$key] = $this->validateType($type, $value);
@@ -67,5 +68,59 @@ class Tuple implements ArrayAccess
     public function toArray(): array
     {
         return $this->data;
+    }
+
+    private function createIterator(array $data): Iterator
+    {
+        return new class($this->types, $data) implements Iterator {
+            /** @var array */
+            private $types;
+
+            /** @var array */
+            private $data;
+
+            /** @var int */
+            private $position;
+
+            public function __construct(array $types, array $data)
+            {
+                $typeCount = count($types);
+
+                $dataCount = count($data);
+
+                if ($typeCount !== $dataCount) {
+                    throw WrongType::fromMessage("Tuple count mismatch, excpected exactly {$typeCount} elements, and got {$dataCount}");
+                }
+
+                $this->types = $types;
+                $this->data = $data;
+                $this->position = 0;
+            }
+
+            public function current(): array
+            {
+                return ['type' => current($this->types), 'value' => current($this->data)];
+            }
+
+            public function next(): void
+            {
+                $this->position++;
+            }
+
+            public function key(): int
+            {
+                return $this->position;
+            }
+
+            public function valid(): bool
+            {
+                return isset($this->types[$this->position]) && array_key_exists($this->position, $this->data);
+            }
+
+            public function rewind(): void
+            {
+                $this->position = 0;
+            }
+        };
     }
 }
